+++ Studenda.Server/Model/Schedule/Subject.cs
@@ -3,6 +3,7 @@ using Studenda.Server.Data.Configuration;
 using Studenda.Server.Model.Common;
 using Studenda.Server.Model.Journal;
 using Studenda.Server.Model.Schedule.Management;
+using Studenda.Server.Model.Security;
 
 namespace Studenda.Server.Model.Schedule;
 
@@ -140,7 +141,7 @@ public class Subject : IdentifiableEntity
     public int? SubjectTypeId { get; set; }
 
     /// <summary>
-    ///     Идентификатор связанного объекта <see cref="Common.Account" />.
+    ///     Идентификатор связанного объекта <see cref="Security.Account" />.
     ///     Необязательное поле.
     /// </summary>
     public int? AccountId { get; set; }
